Updated SSH jump server to import SSH server keys (#6)
* Updated SSH jump server to import SSH server keys from a secret to make them predictable

* Small update to README

* Updated typo

* Fixed typos

* Updated SSH key generation and management

+++ ssh-jump-server/entrypoint.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 
-#Set new keys
-rm -rf /etc/ssh/ssh_host_*
-dpkg-reconfigure openssh-server
+if [[ -d /opt/ssh-server-secrets ]]; then
+    # Keys have been mounted, copy them
+    cp /opt/ssh-server-secrets/* /etc/ssh/
+    chmod 644 /etc/ssh/*.pub
+    for f in $(ls /etc/ssh/ssh_host_*_key); do
+        chmod 600 $f
+    done
+else
+    # Genenerate
+    rm -rf /etc/ssh/ssh_host_*
+    dpkg-reconfigure openssh-server
+fi
 
 #Switch off DNS checking
 sed -i.bak "s/#UseDNS no/UseDNS no/g" /etc/ssh/sshd_config
